+++ WarpEngine.vcxproj.filters
@@ -193,6 +193,10 @@
     <ClInclude Include="src\World\Components.h">
       <Filter>World</Filter>
     </ClInclude>
+    <ClInclude Include="src\World\Components\NametagComponent.h" />
+    <ClInclude Include="src\World\Components\TransformComponent.h" />
+    <ClInclude Include="src\World\Components\CameraComponent.h" />
+    <ClInclude Include="src\Assets\AssetManager.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="vcpkg.json">
