+++ src/Assets/AssetManager.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace Warp
+{
+
+	class AssetManager
+	{
+
+	};
+
+}
