+++ src/World/Components.h
@@ -1,20 +1,5 @@
 #pragma once
 
-#include <string>
-
-namespace Warp
-{
-
-	// Representation of a name of an entity
-	struct NametagComponent
-	{
-		NametagComponent() = default;
-		NametagComponent(std::string_view name)
-			: Nametag(name)
-		{
-		}
-
-		std::string Nametag;
-	};
-
-}
+#include "Components/CameraComponent.h"
+#include "Components/NametagComponent.h"
+#include "Components/TransformComponent.h"
