+++ src/World/Components/CameraComponent.h
@@ -1,0 +1,16 @@
+#pragma once
+
+namespace Warp
+{
+
+	struct CameraComponent
+	{
+
+	};
+
+	struct CameraQuaternionComponent
+	{
+		
+	};
+
+}
